% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change_scene}
\alias{change_scene}
\title{Choose Between Scenes}
\usage{
change_scene(...)
}
\arguments{
\item{...}{One or more \code{shiny_scene} objects.}
}
\value{
A function that processes the request object to delivery a Shiny ui.
}
\description{
Specify a function that uses actions and the request object to choose which
Shiny UI to server.
}
