% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{default_ui}
\alias{default_ui}
\title{Default UI for Unprocessable Requests}
\usage{
default_ui()
}
\value{
A plain text UI with status code 422.
}
\description{
A plain text UI that returns an HTTP status of 422, indicating that the
request was well-formed, but semantically incorrect.
}
\examples{
default_ui()
}
