% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{.compile_methods}
\alias{.compile_methods}
\title{Find Methods Used by Actions}
\usage{
.compile_methods(scenes)
}
\arguments{
\item{scenes}{A list of \code{shiny_scene} objects.}
}
\value{
A character vector of methods accepted by those scenes.
}
\description{
Find Methods Used by Actions
}
\keyword{internal}
