% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.construct_action}
\alias{.construct_action}
\title{Construct a Scene Action}
\usage{
.construct_action(fn_body, negate = FALSE, methods = "GET")
}
\arguments{
\item{fn_body}{The quoted body of the function. This should be everything
that would appear inside the \code{{}} of a function (and the brackets
themselves), inside a call to \code{\link[rlang:expr]{rlang::expr()}}. Usually this will include
variables that should be diffused using \verb{!!}.}

\item{negate}{If \code{TRUE}, trigger the corresponding scene when this action is
\code{not} matched.}

\item{methods}{The http methods which needs to be accepted in order for this
function to make sense. Default "GET" should work in almost all cases.}
}
\value{
A \code{scene_action} object.
}
\description{
Generate the check function for an action, and use it to create a
\code{scene_action} object.
}
\keyword{internal}
