% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.construct_check_fn}
\alias{.construct_check_fn}
\title{Construct a Check Function}
\usage{
.construct_check_fn(body)
}
\arguments{
\item{fn_body}{The quoted body of the function. This should be everything
that would appear inside the \code{{}} of a function (and the brackets
themselves), inside a call to \code{\link[rlang:expr]{rlang::expr()}}. Usually this will include
variables that should be diffused using \verb{!!}.}
}
\value{
A function.
}
\description{
This is a function factory for creating the actual function that will be
called to test the request.
}
\keyword{internal}
