% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{.error_message}
\alias{.error_message}
\title{Generate an Error Message}
\usage{
.error_message(
  parameter,
  parameter_name,
  valid_values,
  special_message,
  level = 2
)
}
\arguments{
\item{parameter}{The argument to test.}

\item{parameter_name}{The argument's name. Eventually this should be
automatically handled through rlang or something, in theory.}

\item{valid_values}{(optional) Expected values of the parameter.}

\item{special_message}{A message tailored to the type of error.}

\item{level}{How deep the check is relative to the original function. Default
= 2.}
}
\description{
Generate an Error Message
}
\keyword{internal}
