% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.new_action}
\alias{.new_action}
\title{Structure a Scene Action}
\usage{
.new_action(check_fn, methods)
}
\arguments{
\item{check_fn}{The function that processes the request to determine if an
associated scene should be returned.}

\item{methods}{The http methods supported by this action.}
}
\value{
A \code{scene_action} object, which is a \code{list} with components \code{check_fn}
and \code{methods}.
}
\description{
Structure a Scene Action
}
\keyword{internal}
