% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{.new_shiny_scene}
\alias{.new_shiny_scene}
\title{Structure a Shiny Scene}
\usage{
.new_shiny_scene(ui, actions)
}
\arguments{
\item{ui}{The ui to return for this set of actions.}

\item{actions}{Zero or more actions required in order to invoke this ui.}
}
\value{
A \code{\link[=shiny_scene-class]{shiny_scene}}.
}
\description{
Structure a Shiny Scene
}
\keyword{internal}
