% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{.parse_ui}
\alias{.parse_ui}
\title{Prepare a Shiny UI for Display}
\usage{
.parse_ui(ui, request)
}
\arguments{
\item{ui}{A 0- or 1-argument function defining the UI of a Shiny app, or a
\code{\link[shiny:reexports]{shiny::tagList()}}.}

\item{request}{The shiny request object.}
}
\value{
A shiny ui as a \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Prepare a Shiny UI for Display
}
\keyword{internal}
