% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action-cookie.R
\name{.req_has_cookie_impl}
\alias{.req_has_cookie_impl}
\title{Check a Request for a Cookie}
\usage{
.req_has_cookie_impl(request, cookie_name, validation_fn, ...)
}
\arguments{
\item{request}{A shiny request object.}

\item{cookie_name}{The cookie that must be present, as a length-1 character
vector.}

\item{validation_fn}{A function that takes the value of the cookie as the
first parameter, and returns \code{TRUE} if the cookie is valid, and \code{FALSE}
otherwise.}

\item{...}{Additional parameters passed on to \code{validation_fn}.}
}
\value{
A length-1 logical vector.
}
\description{
Report whether a request includes a \code{HTTP_COOKIE} object with a specified
\code{cookie_name}, and optionally that the cookie passes
}
\keyword{internal}
