% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action-query.R
\name{.req_has_query_impl}
\alias{.req_has_query_impl}
\title{Check a Request for a Query with a Key}
\usage{
.req_has_query_impl(request, key, values = NULL)
}
\arguments{
\item{request}{A shiny request object.}

\item{key}{The key that must be present, as a length-1 character vector.}

\item{values}{Details about what to look for in the \code{key}. \code{NULL} indicates
that the \code{key} must be present but its contents are unimportant for this
action. Otherwise the actual value of the query must be present in
\code{values}.}
}
\value{
A length-1 logical vector.
}
\description{
Report whether a request includes a \code{QUERY_STRING} object with a specified
\code{key}, and optionally a specific value for that key.
}
\keyword{internal}
