% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action-method.R
\name{.req_uses_method_impl}
\alias{.req_uses_method_impl}
\title{Check a Request for a Method}
\usage{
.req_uses_method_impl(request, method)
}
\arguments{
\item{request}{A shiny request object.}

\item{method}{The expected HTTP method.}
}
\value{
A length-1 logical vector.
}
\description{
Report whether a request includes a \code{REQUEST_METHOD} object with a specified
value.
}
\keyword{internal}
