% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters Used in Various Functions}
\arguments{
\item{request}{A shiny request object.}
}
\description{
Parameters Used in Various Functions
}
\keyword{internal}
