% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{.validate_character_scalar}
\alias{.validate_character_scalar}
\title{Ensure that an Argument is Length-1 Character}
\usage{
.validate_character_scalar(parameter, parameter_name, valid_values)
}
\arguments{
\item{parameter}{The argument to test.}

\item{parameter_name}{The argument's name. Eventually this should be
automatically handled through rlang or something, in theory.}

\item{valid_values}{(optional) Expected values of the parameter.}
}
\value{
The parameter if it is character-scalar.
}
\description{
Ensure that an Argument is Length-1 Character
}
\keyword{internal}
