% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{.validate_values}
\alias{.validate_values}
\title{Ensure that an Argument has Certain Values}
\usage{
.validate_values(parameter, parameter_name, valid_values)
}
\arguments{
\item{parameter}{The argument to test.}

\item{parameter_name}{The argument's name. Eventually this should be
automatically handled through rlang or something, in theory.}

\item{valid_values}{(optional) Expected values of the parameter.}
}
\value{
The parameter.
}
\description{
Ensure that an Argument has Certain Values
}
\keyword{internal}
