% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action-method.R
\name{req_has_method}
\alias{req_has_method}
\alias{req_is_get}
\alias{req_is_post}
\title{Switch Scenes on Method}
\usage{
req_has_method(method, negate = FALSE)

req_is_get(negate = FALSE)

req_is_post(negate = FALSE)
}
\arguments{
\item{method}{The expected HTTP method.}

\item{negate}{If \code{TRUE}, trigger the corresponding scene when this action is
\code{not} matched.}
}
\value{
A \code{scene_action} object, to be used in \code{\link[=set_scene]{set_scene()}}.
}
\description{
Create a \code{scene_action} specifying the HTTP method that must be used (or not
used).
}
\examples{
req_has_method("GET")
req_has_method("POST")
req_is_get()
req_is_get(negate = TRUE)
req_is_post()
req_is_post(negate = TRUE)
}
