% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{scene_action-class}
\alias{scene_action-class}
\alias{scene_action}
\title{\code{scene_action} class}
\description{
A \code{scene_action} object is a \code{list} with components \code{check_fn}
and \code{methods}. It is used to test whether a request should trigger a
particlar scene.
}
\seealso{
\code{\link[=construct_action]{construct_action()}}
}
