% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{shiny_scene-class}
\alias{shiny_scene-class}
\alias{shiny_scene}
\title{\code{shiny_scene} class}
\description{
A \code{shiny_scene} object is a \code{list} with components \code{ui} and
\code{actions}. It is used to define what should display in a Shiny app in
different scenarios.
}
\seealso{
\code{\link[=set_scene]{set_scene()}}
}
