% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenes-package.R
\docType{package}
\name{scenes-package}
\alias{scenes}
\alias{scenes-package}
\title{scenes: Switch Between Alternative 'shiny' UIs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Sometimes it is useful to serve up alternative 'shiny' UIs depending on information passed in the request object, such as the value of a cookie. This packages facilitates such switches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/scenes}
  \item \url{https://r4ds.github.io/scenes/}
  \item Report bugs at \url{https://github.com/r4ds/scenes/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
