% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_scene}
\alias{set_scene}
\title{Link a UI to Required Actions}
\usage{
set_scene(ui, ...)
}
\arguments{
\item{ui}{A shiny ui.}

\item{...}{One or more \code{scene_action} objects.}
}
\value{
A \code{shiny_scene} object, which is a list with components \code{ui} and
\code{actions}.
}
\description{
A scene is a shiny ui and the actions that trigger it.
}
