% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_scene}
\alias{set_scene}
\title{Link a UI to Required Actions}
\usage{
set_scene(ui, ...)
}
\arguments{
\item{ui}{A shiny ui.}

\item{...}{Zero or more \code{\link[=scene_action-class]{scene_actions}}.}
}
\value{
A \code{\link[=shiny_scene-class]{shiny_scene}}.
}
\description{
Define a \code{\link[=shiny_scene-class]{shiny_scene}} by linking a UI to zero or more
\code{\link[=scene_action-class]{scene_action}} requirements.
}
\examples{
scene1 <- set_scene(
  "A shiny ui",
  req_has_query("scene", 1)
)
scene1
scene2 <- set_scene(
  "Another shiny ui",
  req_has_query("scene", 2)
)
scene2
}
